Added Everything xdxdxd
This project has a lot of changes. Was so much things but finally i found what i want to do with it. The game will be a Gradius. 

In this commit we add almost everything, like the shoot, and other basic stuff. The project still green, but is something right now. :3